don't create double doors
dynamicCards was reset every time setup() was executed, so live programming would recreate doors. This could have been avoided by not resetting dynamicCards if it exists.

But more importantly, persistence only preserves card data, not other properties of actors. This means dynamicCards was always empty after starting from persisted data (as opposed to a snapshot).

The new mechanism now looks at the names of all cards, which are preserved as part of the card data.